refactor: move 404 page to pages
